+++ app/controller/admin/BlogPostController.php
@@ -3,8 +3,7 @@
 	namespace App\Controller\Admin;
 
 	use App\Controller\AppController;
-	use App\Entity;
-	use App\FileUploadHelper;
+	use Core\Helper\FileUploadHelper;
 
 	class BlogPostController extends AppController {
 		public function __construct() {
