+++ app/entity/BlogPostEntity.php
@@ -2,6 +2,8 @@
 
 	namespace App\Entity;
 
+	use \Core\Entity\Entity;
+
 	class BlogPostEntity extends Entity {
 		/*
 		 * Gets url
