+++ composer.json
@@ -15,7 +15,8 @@
     ],
 	"autoload": {
         "psr-4": {
-            "App\\": "app"
+            "App\\": "app",
+            "Core\\": "core"
         }
     }
 }
