+++ core/Config.php
@@ -1,6 +1,6 @@
 <?php
 
-	namespace App;
+	namespace Core;
 
 	class Config {
 		private $settings = [];
