+++ core/Session.php
@@ -1,6 +1,6 @@
 <?php
 
-	namespace App;
+	namespace Core;
 
 	class Session {
 		public function __construct() {
