+++ core/database/MySQLDatabase.php
@@ -1,6 +1,6 @@
 <?php
 
-	namespace App\Database;
+	namespace Core\Database;
 
 	use \PDO;
