+++ core/helper/FileUploadHelper.php
@@ -1,6 +1,6 @@
 <?php
 
-	namespace App;
+	namespace Core\Helper;
 
 	class FileUploadHelper {
 		public $file;
