+++ core/router/Route.php
@@ -1,6 +1,6 @@
 <?php
 	
-	namespace App\Router;
+	namespace Core\Router;
 
 	class Route {
 		private $path;
